Update VoiceEndCallRequest to use float for duration_seconds and log … (#3549)
…conversation details in end_voice_call

<!-- PR title must include a Notion task ID: [PAL-XXXX] Your title here
-->

## Summary

<!-- Describe your changes -->


## Test Plan

<!-- How were these changes tested? -->

+++ api/schemas/internal/voice_init.py
@@ -51,7 +51,7 @@ class VoiceEndCallRequest(BaseModel):
     call_id: str = Field(..., description="SIP call identifier")
     caller_number: str = Field(..., description="Customer phone number")
     dialed_number: str = Field(..., description="Business phone number")
-    duration_seconds: int = Field(..., description="Call duration in seconds")
+    duration_seconds: float = Field(..., description="Call duration in seconds")
     conversation: list[dict[str, Any]] = Field(
         ..., description="Call conversation history"
     )
